requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.5.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.17
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==1.11.13
django-pure-pagination==0.3.0
gevent==1.3.2.post0
greenlet==0.4.13
idna==2.6
kombu==3.0.37
lxml==4.2.3
Markdown==2.6.11
Pillow==5.1.0
pycharts==0.1.5
pycparser==2.18
pygame==1.9.3
pymongo==3.6.1
PyMySQL==0.7.11
pytz==2018.4
qqbot==2.3.10
requests==2.18.4
selenium==3.13.0
six==1.11.0
tornado==4.5.2
tzlocal==1.5.1
urllib3==1.22
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
wxPython==4.0.1
